Add destructive cell deletion confirmation (#5664)
Fixes #5658

This change adds a keymap configuration (`destructive_delete`, default
**True**) to enable controlled, destructive deletion of cells via
keyboard shortcut. Video:

- First deletion is "cheap" (`d = 10`), and automatically deleted
- Section deletion is "expensive", and adds confirmation step

<img width=500
src="https://github.com/user-attachments/assets/ddbe0836-c5ef-4b6c-b531-ac4427a27461"/>


The idea is to allow deletion of cells, but to have a confirmation step
(i.e., "are you sure you want to do that?") when we consider a deletion
_could_ be "expensive". The definition of expensive is up for question
but right now we can look at:

- Whether a cells execution time is above some threshold (e.g., 2s)
- Whether a cell declares variables used by other cells

If neither of those conditions are met, we automatically delete the
cell.

What I hope this unlocks is the ability to readily delete cells for the
common editing use case, and give pause for the moments where a deletion
might be costly or unintentional. My preference is to have opt-out
unless the user explicitly prefers stricter safeguards.

---------

+++ frontend/src/components/editor/cell/code/cell-editor.tsx
@@ -10,6 +10,7 @@ import { DelayMount } from "@/components/utils/delay-mount";
 import { aiCompletionCellAtom } from "@/core/ai/state";
 import { maybeAddMarimoImport } from "@/core/cells/add-missing-import";
 import { useCellActions } from "@/core/cells/cells";
+import { usePendingDeleteService } from "@/core/cells/pending-delete-service";
 import type { CellData, CellRuntimeState } from "@/core/cells/types";
 import { setupCodeMirror } from "@/core/codemirror/cm";
 import {
@@ -91,6 +92,7 @@ const CellEditorInternal = ({
   const [aiCompletionCell, setAiCompletionCell] = useAtom(aiCompletionCellAtom);
   const deleteCell = useDeleteCellCallback();
   const { saveOrNameNotebook } = useSaveNotebook();
+  const pendingDeleteService = usePendingDeleteService();
 
   const loading = status === "running" || status === "queued";
   const cellActions = useCellActions();
@@ -102,6 +104,11 @@ const CellEditorInternal = ({
     // Cannot delete running cells, since we're waiting for their output.
     if (loading) {
       return false;
+    }
+
+    if (pendingDeleteService.idle && userConfig.keymap.destructive_delete) {
+      pendingDeleteService.submit([cellId]);
+      return true;
     }
 
     deleteCell({ cellId });
